Ticket #13144 - Contracting Parties Counter
Deleted MERC_AccountDesignatedIndividualCounter and moved code to
MERC_AccountRollupHandler. Added logic go count up the number of
Contracting parties as well as the number of Designated individuals.
The initial logic for Designated Individuals was actually wrong as I
incorrectly assumed that a Designated Individual was the only thing
returned in a subquery ‘select Id from Designated_Individuals__r’. This
relationship has been used for other lookups as well (such as
contracting party) and the distinguish factor is record type.

Updated the account trigger to use the new class.