Ticket #13232 - Manage Meeting Days
1. when a new day is added we insert it right away so that hotel
allocations can be created for any hotel already existing
2. when a hotel is added, we NO LONGER insert it right away. when the
user searches for the hotel and selects it the save hotels button will
be "clicked" and then the hotel will be saved. At that point the page
will rerender with the new meeting days added.
3. I have added the IE9 hack to this page to get the rerenders to work
in IE9 regardless of compatibility mode.
4. SEARCH FUNCTIONALITY CHANGED: this is a bit important as it may be a
business rule issue. I have removed the leading wildcard from the hotel
typeahead search. The leading wildcard would result in a query on
"marriott" searching for "%marriott%". While account name is indexed,
slapping a wildcard at the beginning effectively throws that index out
the window, resulting in long wait times for the query. (When initially
tested in GSO1FULL I encountered actual timeouts).
5. updated the rerenders to allow messages to show up in appropriate
places and clear out when certain actions are taken, just a general
improvement to user experience.