Ticket #13251 - CST search
updated the controller to check for a record type selection and throw
error if none. also added a new field CST_Validation_Bypass_MERC__c
that allows this page to bypass validation rules that are preventing
the page from saving.

on the vf page moved the display of “participant added successfully” to
the handleAddResponse function so it only showed if there was no error.