Ticket #13277 - Adding Meeting Participants
Added a required fields method that pulls information out of the
Add_Participant_Rec_Type_Req_Fields_MERC__c custom setting based on the
current record type selected and then builds a list of required fields
to be displayed on the page in order to add a meeting participant with
that record type.

Required fields are sourced from fields sets and field set name to
reference is configured on the custom setting.