Ticket #13491 - Convert MP val rule to trigger
Added new Fieldset: Meeting_Participant_MERC__c.ToV_Validation_Fields
to hold the fields that need to be validated under this trigger

Created
:: Custom Label: MERC_MeetingParticipantTovValidation
:: MERC_MeetingParticipantTovValidation
:: MERC_MeetingParticipantTovValidationTest

MERC_MeetingParticipantTovValidation has a static boolean
tovValidationByPass that defaults to false and updates to true after
first pass. this boolean can be set from other classes to skip the
confirm validation, especially when it is causing problems due to
rollups.