Updated to fix calculation bug
The value for the firstDay and lastDay was being set by pulling the
travel itinerary Segment Begin and Segment End dates and calling the
.date() method on them. this posed a problem if the locale of the user
was behind GMT as the date returned would be 1 day behind the actual
date. e.g. if the date was ‘2014-06-02 00:00:00’ then calling .date()
on that returned 2014-06-01.