Ticket #13711 - Update account merge
Updated the merge accounts page to function under the new understanding
that the integration will perform a salesforce merge and the losing
account will no longer exist. the page looks for the Customer Id that
was stamped to Prior_Account_Id_GLBL__c and uses that to determine if
the related item appears in the winning or losing list.

merging no longer requires reparenting as that is already done for us.
it simply removes the old customer id.