Ticket #13726 - Team Member Shares
created a utility with two methods that take a json serialized string
representing the the team member share records to delete or create. the
trigger handler has been update to call these @future methods and pass
in the json serialized string.

the class is without sharing so this effectively runs as